requirements.txt
PyQt5==5.15.2
ffmpeg-python

send2trash